[Rgen] Share DataModel between the generator and the transformer.
In this change we are making the necesary changes to share the data
model code between the generator and the transformer.

When we talk about sharing code in C# the average dotnet developer
would create a csproj with the shared code and would reference it. In
normal circumtances this is not a bad option, but our case is different
and here are some of the reasons we cannot take that path:

1. Roslyn code generators with more than one dll are problematic to
   distribute. You can already see how bad it is by looking at the dll
   shared between the code generator and the analyzer. The shared
   library has to be added as an analyzer so that it loads with the
   generator. We want to keep the number of dlls to a minimun.
2. While the data models are similar, they are not identical. The
   attributes used by the code generator and the old api defintions are
   different. We would around this by splitting the definitions in 2
   files:
   a - A common file with all date that does not depend on the
   attributes.
   b - A application specific file that contains the properties needed
   by the common parts BUT that uses structures specific for the
   application. An example is the ExportData. The ExportData is a
   generic in the generator (which contains flags) while it is not in
   the transformer. We also split the TryCreate and Constructor methods
   since those need to init the application specific parths.

In this commit we do not implement the transformer parts since that
would make the diff hader. We are only moving code around and adding
dummy implementations for the transformer project to compile.
