+++ src/rgen/Microsoft.Macios.Generator/DataModel/EnumMember.cs
@@ -2,6 +2,8 @@
 // Licensed under the MIT License.
 using System;
 using System.Collections.Immutable;
+using System.Reflection;
+using System.Runtime.InteropServices;
 using System.Text;
 using Microsoft.Macios.Generator.Attributes;
 using Microsoft.Macios.Generator.Availability;
@@ -13,7 +15,8 @@ namespace Microsoft.Macios.Generator.DataModel;
 /// Structure that represents a change that was made by the user on enum members that has to be
 /// reflected in the generated code.
 /// </summary>
-readonly struct EnumMember : IEquatable<EnumMember> {
+[StructLayout(LayoutKind.Auto)]
+readonly partial struct EnumMember : IEquatable<EnumMember> {
 	/// <summary>
 	/// Get the name of the member.
 	/// </summary>
@@ -25,37 +28,10 @@ readonly struct EnumMember : IEquatable<EnumMember> {
 	public SymbolAvailability SymbolAvailability { get; }
 
 	/// <summary>
-	/// The data of the field attribute used to mark the value as a binding.
-	/// </summary>
-	public FieldInfo<EnumValue>? FieldInfo { get; }
-
-	/// <summary>
 	/// Get the attributes added to the member.
 	/// </summary>
 	public ImmutableArray<AttributeCodeChange> Attributes { get; }
-
-	/// <summary>
-	/// Create a new change that happened on a member.
-	/// </summary>
-	/// <param name="name">The name of the changed member.</param>
-	/// <param name="libraryName">The library name of the smart enum.</param>
-	/// <param name="libraryPath">The library path to the library, null if it is a known frameworl.</param>
-	/// <param name="fieldData">The binding data attached to this enum value.</param>
-	/// <param name="symbolAvailability">The symbol availability of the member.</param>
-	/// <param name="attributes">The list of attribute changes in the member.</param>
-	public EnumMember (string name,
-		string libraryName,
-		string? libraryPath,
-		FieldData<EnumValue>? fieldData,
-		SymbolAvailability symbolAvailability,
-		ImmutableArray<AttributeCodeChange> attributes)
-	{
-		Name = name;
-		FieldInfo = fieldData is null ? null : new (fieldData.Value, libraryName, libraryPath);
-		SymbolAvailability = symbolAvailability;
-		Attributes = attributes;
-	}
-
+	
 	/// <summary>
 	/// Create a new change that happened on a member.
 	/// </summary>
