+++ src/rgen/Microsoft.Macios.Generator/DataModel/TypeInfo.cs
@@ -5,14 +5,13 @@ using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Text;
 using Microsoft.CodeAnalysis;
-using Microsoft.Macios.Generator.Extensions;
 
 namespace Microsoft.Macios.Generator.DataModel;
 
 /// <summary>
 /// Readonly structure that represents a change in a method return type.
 /// </summary>
-readonly struct TypeInfo : IEquatable<TypeInfo> {
+readonly partial struct TypeInfo : IEquatable<TypeInfo> {
 
 	/// <summary>
 	/// Type of the parameter.
@@ -135,50 +134,7 @@ readonly struct TypeInfo : IEquatable<TypeInfo> {
 		IsArray = isArray;
 		IsReferenceType = isReferenceType;
 	}
-
-	internal TypeInfo (ITypeSymbol symbol) :
-		this (
-			symbol is IArrayTypeSymbol arrayTypeSymbol
-				? arrayTypeSymbol.ElementType.ToDisplayString ()
-				: symbol.ToDisplayString ().Trim ('?', '[', ']'),
-			symbol.SpecialType)
-	{
-		IsNullable = symbol.NullableAnnotation == NullableAnnotation.Annotated;
-		IsBlittable = symbol.IsBlittable ();
-		IsSmartEnum = symbol.IsSmartEnum ();
-		IsArray = symbol is IArrayTypeSymbol;
-		IsReferenceType = symbol.IsReferenceType;
-		IsInterface = symbol.TypeKind == TypeKind.Interface;
-		IsNativeIntegerType = symbol.IsNativeIntegerType;
-		IsNativeEnum = symbol.HasAttribute (AttributesNames.NativeEnumAttribute);
-
-		// data that we can get from the symbol without being INamedType
-		symbol.GetInheritance (
-			isNSObject: out isNSObject,
-			isNativeObject: out isINativeObject,
-			parents: out parents,
-			interfaces: out interfaces);
-
-		// try to get the named type symbol to have more educated decisions
-		var namedTypeSymbol = symbol as INamedTypeSymbol;
-
-		// store the enum special type, useful when generate code that needs to cast
-		EnumUnderlyingType = namedTypeSymbol?.EnumUnderlyingType?.SpecialType;
-
-		if (!IsReferenceType && IsNullable && namedTypeSymbol is not null) {
-			// get the type argument for nullable, which we know is the data that was boxed and use it to 
-			// overwrite the SpecialType 
-			var typeArgument = namedTypeSymbol.TypeArguments [0];
-			SpecialType = typeArgument.SpecialType;
-			MetadataName = SpecialType is SpecialType.None or SpecialType.System_Void
-				? null : typeArgument.MetadataName;
-		} else {
-			MetadataName = SpecialType is SpecialType.None or SpecialType.System_Void
-				? null : symbol.MetadataName;
-		}
-
-	}
-
+	
 	/// <inheritdoc/>
 	public bool Equals (TypeInfo other)
 	{
