+++ src/rgen/Microsoft.Macios.Transformer/DataModel/CodeChanges.Transformer.cs
@@ -1,0 +1,22 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Microsoft.Macios.Generator.DataModel;
+
+readonly partial struct CodeChanges {
+
+	/// <summary>
+	/// Represents the type of binding that the code changes are for.
+	/// </summary>
+	public BindingType BindingType => throw new NotImplementedException ();
+	
+	public BindingInfo BindingInfo => throw new NotImplementedException ();
+	
+	public CodeChanges ()
+	{
+		FullyQualifiedSymbol = "";
+		IsStatic = false;
+		IsPartial = false;
+		IsAbstract = false;
+	}
+}
