+++ src/rgen/Microsoft.Macios.Transformer/DataModel/Method.Transformer.cs
@@ -1,0 +1,39 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Collections.Immutable;
+using Microsoft.CodeAnalysis;
+using Microsoft.Macios.Generator.Availability;
+using Microsoft.Macios.Transformer.Attributes;
+
+namespace Microsoft.Macios.Generator.DataModel;
+
+readonly partial struct Method {
+	
+	/// <summary>
+	/// The data of the export attribute used to mark the value as a property binding. 
+	/// </summary>
+	public ExportData ExportMethodData { get; }
+
+	/// <summary>
+	/// True if the method was exported with the MarshalNativeExceptions flag allowing it to support native exceptions.
+	/// </summary>
+	public bool MarshalNativeExceptions => throw new NotImplementedException ();
+	
+	public Method (string type, string name, TypeInfo returnType,
+		SymbolAvailability symbolAvailability,
+		ExportData exportMethodData,
+		ImmutableArray<AttributeCodeChange> attributes,
+		ImmutableArray<SyntaxToken> modifiers,
+		ImmutableArray<Parameter> parameters)
+	{
+		Type = type;
+		Name = name;
+		ReturnType = returnType;
+		SymbolAvailability = symbolAvailability;
+		ExportMethodData = exportMethodData;
+		Attributes = attributes;
+		Modifiers = modifiers;
+		Parameters = parameters;
+	}
+}
