+++ src/rgen/Microsoft.Macios.Transformer/Extensions/TypeSymbolExtensions.Transformer.cs
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using Microsoft.CodeAnalysis;
+
+namespace Microsoft.Macios.Generator.Extensions;
+
+static partial class TypeSymbolExtensions {
+	
+	public static bool IsSmartEnum (this ITypeSymbol symbol)
+	{
+		throw new NotImplementedException ();
+	}
+}
