+++ src/rgen/Microsoft.Macios.Transformer/Microsoft.Macios.Transformer.csproj
@@ -25,6 +25,9 @@
         <Compile Include="../Microsoft.Macios.Generator/DictionaryComparer.cs" >
             <Link>DictionaryComparer.cs</Link>
         </Compile>
+        <Compile Include="../Microsoft.Macios.Generator/CollectionComparer.cs" >
+            <Link>CollectionComparer.cs</Link>
+        </Compile>
         <Compile Include="../Microsoft.Macios.Generator/Attributes/ObsoletedOSPlatformData.cs" >
             <Link>Attributes/ObsoletedOSPlatformData.cs</Link>
         </Compile>
@@ -37,12 +40,28 @@
         <Compile Include="../Microsoft.Macios.Generator/Availability/*.cs" >
             <Link>Availability/*.cs</Link>
         </Compile>
+        <Compile Include="../Microsoft.Macios.Generator/Extensions/ParameterSyntaxExtensions.cs" >
+            <Link>Extensions/ParameterSyntaxExtensions.cs</Link>
+        </Compile>
         <Compile Include="../Microsoft.Macios.Generator/Extensions/StringExtensions.cs" >
             <Link>Extensions/StringExtensions.cs</Link>
         </Compile>
         <Compile Include="../Microsoft.Macios.Generator/Extensions/TypeSymbolExtensions.Core.cs" >
             <Link>Extensions/TypeSymbolExtensions.Core.cs</Link>
         </Compile>
+    </ItemGroup>
+
+    <ItemGroup>
+        <Compile Include="../Microsoft.Macios.Generator/DataModel/*.cs" >
+            <Link>DataModel/*.cs</Link>
+        </Compile>
+        <!-- Remove those files that are just used by the generator -->
+        <None Remove="../Microsoft.Macios.Generator/DataModel/BindingInfo.cs" />
+        <Compile Remove="../Microsoft.Macios.Generator/DataModel/BindingInfo.cs" />
+        <None Remove="../Microsoft.Macios.Generator/DataModel/FieldInfo.cs" />
+        <Compile Remove="../Microsoft.Macios.Generator/DataModel/FieldInfo.cs" />
+        <None Remove="../Microsoft.Macios.Generator/DataModel/*.Generator.cs" />
+        <Compile Remove="../Microsoft.Macios.Generator/DataModel/*.Generator.cs" />
     </ItemGroup>
 
     <ItemGroup>
